requirements-dev.txt
pytest
bump2version
python-dotenv
pickle